Create string_refinement_util for utility function
string_refinement.cpp file is becoming too large.
This groups utility classes used by string_refinement in new files to
reduce string_refinement.cpp a bit.
