+++ src/solvers/refinement/string_refinement.h
@@ -27,56 +27,10 @@ Author: Alberto Griggio, alberto.griggio@gmail.com
 #include <solvers/refinement/string_constraint.h>
 #include <solvers/refinement/string_constraint_generator.h>
 #include <solvers/refinement/string_refinement_invariant.h>
+#include <solvers/refinement/string_refinement_util.h>
 
 #define DEFAULT_MAX_NB_REFINEMENT std::numeric_limits<size_t>::max()
 #define CHARACTER_FOR_UNKNOWN '?'
-
-struct index_set_pairt
-{
-  std::map<exprt, std::set<exprt>> cumulative;
-  std::map<exprt, std::set<exprt>> current;
-};
-
-struct string_axiomst
-{
-  std::vector<string_constraintt> universal;
-  std::vector<string_not_contains_constraintt> not_contains;
-};
-
-/// Represents arrays of the form `array_of(x) with {i:=a} with {j:=b} ...`
-/// by a default value `x` and a list of entries `(i,a)`, `(j,b)` etc.
-class sparse_arrayt
-{
-public:
-  /// Initialize a sparse array from an expression of the form
-  /// `array_of(x) with {i:=a} with {j:=b} ...`
-  /// This is the form in which array valuations coming from the underlying
-  /// solver are given.
-  explicit sparse_arrayt(const with_exprt &expr);
-
-  /// Creates an if_expr corresponding to the result of accessing the array
-  /// at the given index
-  virtual exprt to_if_expression(const exprt &index) const;
-
-protected:
-  exprt default_value;
-  std::vector<std::pair<std::size_t, exprt>> entries;
-};
-
-/// Represents arrays by the indexes up to which the value remains the same.
-/// For instance `{'a', 'a', 'a', 'b', 'b', 'c'}` would be represented by
-/// by ('a', 2) ; ('b', 4), ('c', 5).
-/// This is particularly useful when the array is constant on intervals.
-class interval_sparse_arrayt final : public sparse_arrayt
-{
-public:
-  /// An expression of the form `array_of(x) with {i:=a} with {j:=b}` is
-  /// converted to an array `arr` where for all index `k` smaller than `i`,
-  /// `arr[k]` is `a`, for all index between `i` (exclusive) and `j` it is `b`
-  /// and for the others it is `x`.
-  explicit interval_sparse_arrayt(const with_exprt &expr);
-  exprt to_if_expression(const exprt &index) const override;
-};
 
 class string_refinementt final: public bv_refinementt
 {
